Renaming and relocating Widget components: message peuces
