+++ widget/components/chat/msgs/pieces/AttachmentMsgPiece.vue
