+++ widget/components/chat/msgs/pieces/FileUploadMsgPiece.vue
