+++ widget/components/chat/msgs/pieces/ReferencesMsgPiece.vue
