+++ widget/components/chat/msgs/pieces/TextMsgPiece.vue
